Removed last traces of frontend
