+++ app.js
@@ -7,21 +7,18 @@ const env = require('dotenv').config() // eslint-disable-line
 const app = express()
 const router = express.Router() // eslint-disable-line
 const jsonParser = bodyParser.json() // eslint-disable-line
+const port = 6789
 
-const routes = require('./routes/index')
-const api = require('./routes/api')
-const backend = require('./routes/backend')
+const index = require('./routes/index')
+const dashboard = require('./routes/dashboard')
 
-app.set('views', path.join(__dirname, 'views'))
-app.set('view engine', 'pug')
 app.locals.pretty = true
 
 app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: false }))
 
-app.use('/', routes)
-app.use('/api', api)
-app.use('/backend', backend)
+app.use('/', index)
+app.use('/dashboard', dashboard)
 
 app.use((req, res, next) => {
   let err = new Error('Not Found')
@@ -31,20 +28,14 @@ app.use((req, res, next) => {
 
 app.use((err, req, res, next) => {
   res.status(err.status || 500)
-  if (app.get('env') !== 'development') {
-    err = ''
-  }
-  res.render('error', {
-    message: err.message,
-    error: err
-  })
+  res.send(err)
 })
 
-app.listen(6789, () => {
+app.listen(port, () => {
   console.log(
-        ' Online and logged on at Port 6789\n',
-        `NODE_ENV set to ${process.env.NODE_ENV} mode\n`
-    )
+    `Online and logged on at Port ${port}\n`, 
+    `NODE_ENV set to ${process.env.NODE_ENV} mode\n`
+  )
 })
 
 module.exports = app
